Refactor reasoner test package structure
