+++ test/behaviour/resolution/reasoner/attribute_attachment/AttributeAttachmentTest.java
@@ -32,7 +32,7 @@ import java.lang.reflect.InvocationTargetException;
         strict = true,
         plugin = "pretty",
         glue = "grakn.core.test.behaviour",
-        features = "external/graknlabs_verification/behaviour/graql/reasoner/resolution/attribute-attachment.feature",
+        features = "external/graknlabs_verification/behaviour/graql/reasoner/attribute-attachment.feature",
         tags = "not @ignore and not @ignore-grakn-core"
 )
 public class AttributeAttachmentTest {
