+++ test/behaviour/resolution/reasoner/relation_inference/RelationInferenceTest.java
@@ -32,7 +32,7 @@ import java.lang.reflect.InvocationTargetException;
         strict = true,
         plugin = "pretty",
         glue = "grakn.core.test.behaviour",
-        features = "external/graknlabs_verification/behaviour/graql/reasoner/resolution/relation-inference.feature",
+        features = "external/graknlabs_verification/behaviour/graql/reasoner/relation-inference.feature",
         tags = "not @ignore and not @ignore-grakn-core"
 )
 public class RelationInferenceTest {
