+++ test/behaviour/resolution/reasoner/roleplayer_attachment/RoleplayerAttachmentTest.java
@@ -32,7 +32,7 @@ import java.lang.reflect.InvocationTargetException;
         strict = true,
         plugin = "pretty",
         glue = "grakn.core.test.behaviour",
-        features = "external/graknlabs_verification/behaviour/graql/reasoner/resolution/roleplayer-attachment.feature",
+        features = "external/graknlabs_verification/behaviour/graql/reasoner/roleplayer-attachment.feature",
         tags = "not @ignore and not @ignore-grakn-core"
 )
 public class RoleplayerAttachmentTest {
