+++ test/behaviour/resolution/reasoner/value_predicate/ValuePredicateTest.java
@@ -32,7 +32,7 @@ import java.lang.reflect.InvocationTargetException;
         strict = true,
         plugin = "pretty",
         glue = "grakn.core.test.behaviour",
-        features = "external/graknlabs_verification/behaviour/graql/reasoner/resolution/value-predicate.feature",
+        features = "external/graknlabs_verification/behaviour/graql/reasoner/value-predicate.feature",
         tags = "not @ignore and not @ignore-grakn-core"
 )
 public class ValuePredicateTest {
