+++ test/behaviour/resolution/reasoner/variable_role/BUILD
@@ -38,7 +38,7 @@ java_test(
         "//test/behaviour/resolution:steps",
     ],
     data = [
-        "@graknlabs_verification//behaviour/graql/reasoner/resolution:variable-role.feature",
+        "@graknlabs_verification//behaviour/graql/reasoner:variable-role.feature",
     ],
     classpath_resources = ["//test/resources:logback-test"],
     size = "medium",
