+++ test/behaviour/resolution/reasoner/variable_role/VariableRoleTest.java
@@ -32,7 +32,7 @@ import java.lang.reflect.InvocationTargetException;
         strict = true,
         plugin = "pretty",
         glue = "grakn.core.test.behaviour",
-        features = "external/graknlabs_verification/behaviour/graql/reasoner/resolution/variable-role.feature",
+        features = "external/graknlabs_verification/behaviour/graql/reasoner/variable-role.feature",
         tags = "not @ignore and not @ignore-grakn-core"
 )
 public class VariableRoleTest {
